# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "The REST API for Connector integration enables you to interact with\
    \ connectors for the cloud stores. It will enable customer to reuse the content\
    \ they’ve already created,manage centrally in CEC."
  version: "2019.01.01"
  title: "REST API for Connector Integration"
  x-summary: "The REST API for Connector integration enables you to interact with\
    \ connectors for the cloud stores. It will enable customer to reuse the content\
    \ they’ve already created,manage centrally in CEC."
...
schemes:
- "https"
- "http"
consumes:
- "application/json"
produces:
- "application/json"
- "text/html"
paths:
...
  /api/{version}/authorization/authorizationURLs:
    post:
      tags:
      - "AuthorizationResource"
      summary: "Create an authorization URL and related information. This API will\
        \ need to be implemented if remote store suppots oauth style of authorization"
      operationId: "apiVersionAuthorizationAuthorizationURLsPOST"
      produces:
      - "application/json"
      parameters:
      - name: "X-CEC-clientId"
        in: "header"
        description: "Connector Framework will pass all the custom properties and\
          \ proxy values via Http headers."
        required: true
        type: "string"
      - name: "X-CEC-clientSecret"
        in: "header"
        description: "Version value of the API for example the current release version\
          \ is v1"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version value of the API for example the current release version\
          \ is v1"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "The request body defines the details of the request."
        required: false
        schema:
          $ref: "#/definitions/CreateAuthorizationRequestInfo"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/CreateAuthorizationResponseInfo"
      x-swagger-router-controller: "AuthorizationResource"
  /api/{version}/authorization/completedAuthorizations:
    post:
      tags:
      - "AuthorizationResource"
      summary: "To complete the oauth based authorization. This API will need to be\
        \ implemented if remote store suppots oauth style of authorization"
      operationId: "apiVersionAuthorizationCompletedAuthorizationsPOST"
      produces:
      - "application/json"
      parameters:
      - name: "X-CEC-clientId"
        in: "header"
        description: "Client Id for the application"
        required: true
        type: "string"
      - name: "X-CEC-clientSecret"
        in: "header"
        description: "Client Secret for the application"
        required: true
        type: "string"
      - name: "X-CEC-code"
        in: "header"
        description: "Access Code for the application"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version value of the API for example the current release version\
          \ is v1"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "The request body defines the details of the request."
        required: true
        schema:
          $ref: "#/definitions/CompleteAuthorizationRequestInfo"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/CompleteAuthorizationResponseInfo"
      x-swagger-router-controller: "AuthorizationResource"
  /api/{version}/authorization/basicAuthorizations:
    post:
      tags:
      - "AuthorizationResource"
      summary: "To verify the credential for basic authorization. This API will need\
        \ to be implemented if remote store suppots Basic authentication"
      operationId: "apiVersionAuthorizationBasicAuthorizationsPOST"
      produces:
      - "application/json"
      parameters:
      - name: "headers"
        in: "header"
        description: "Connector Framework will pass all the custom properties and\
          \ proxy values via Http headers."
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version value of the API for example the current release version\
          \ is v1"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation."
      x-swagger-router-controller: "AuthorizationResource"
  /api/{version}/content:
    get:
      tags:
      - "ContentResource"
      summary: "Get the content of a filesystem object.The response entity is the\
        \ content of the file. The response should include the following headers:\
        \ Content-Length -> set to the entity length or -1 if unknown, Content-Type\
        \ -> set to the content type or <code>application/octet-stream</code> if a\
        \ more accurate type is not possible, Content-Disposition -> set to disposition\
        \ type of attachment and include a filename parameter (for example,Content-Disposition:\
        \ attachment; filename=''meeting agenda.doc''). Note that file names with\
        \ spaces must be quoted. See RFC 6266 for details about the format and encoding\
        \ of the header."
      operationId: "apiVersionContentGET"
      produces:
      - "*/*"
      parameters:
      - name: "uri"
        in: "query"
        description: "URI of the filesystem object default value if empty is root"
        required: false
        type: "string"
      - name: "version"
        in: "path"
        description: "Version value of the API for example the current release version\
          \ is v1"
        required: true
        type: "string"
      - name: "X-CEC-accessToken"
        in: "header"
        description: "Access token"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation. It will return an InputStream for reading\
            \ the bytes of the file"
        403:
          description: "FILESYSTEM_OBJECT_NOT_FOUND or OAUTH_NOT_AUTHORIZED"
      x-swagger-router-controller: "ContentResource"
  /api/{version}/content/getthumbnail:
    get:
      tags:
      - "ContentResource"
      summary: "Get the content of a thumbnail of the file.The response entity is\
        \ the content of the thumbnail. The response should include the following\
        \ headers: Content-Length -> set to the entity length or -1 if unknown, Content-Type\
        \ -> set to the content type or <code>application/octet-stream</code> if a\
        \ more accurate type is not possible, Content-Disposition -> set to disposition\
        \ type of attachment and include a filename parameter (for example,Content-Disposition:\
        \ attachment; filename=''meeting agenda.doc''). Note that file names with\
        \ spaces must be quoted. See RFC 6266 for details about the format and encoding\
        \ of the header."
      operationId: "apiVersionContentGetthumbnailGET"
      produces:
      - "*/*"
      parameters:
      - name: "uri"
        in: "query"
        description: "URI of the filesystem object. The object's unique identifier\
          \ from a string representation."
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version value of the API for example the current release version\
          \ is v1"
        required: true
        type: "string"
      - name: "headers"
        in: "header"
        description: "Connector Framework will pass all the custom properties and\
          \ proxy values via Http headers."
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation. It will return an InputStream for reading\
            \ the bytes of the thumbnail"
        403:
          description: "FILESYSTEM_OBJECT_NOT_FOUND or OAUTH_NOT_AUTHORIZED"
      x-swagger-router-controller: "ContentResource"
...
  /api/{version}/filesystem:
    post:
      tags:
      - "FilesystemResource"
      summary: "Return meta data about an object in remote files system"
      operationId: "apiVersionFilesystemPOST"
      produces:
      - "application/json"
      parameters:
      - name: "headers"
        in: "header"
        description: "Connector Framework will pass all the custom properties and\
          \ proxy values via Http headers."
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version value of the API for example the current release version\
          \ is v1"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "The request body defines the details of the request."
        required: true
        schema:
          $ref: "#/definitions/FileSystemRequestInfo"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/FileSystemResponseInfo"
      x-swagger-router-controller: "FilesystemResource"
definitions:
...
  CreateAuthorizationRequestInfo:
    type: "object"
    properties:
      redirectURL:
        type: "string"
        description: "The redirect URL to which the authorization service should redirect\
          \ when finished. For example, the URL of a specific page served by the CEC\
          \ server."
    description: "Authorization-request creation info."
    example:
      redirectURL: "redirectURL"
  CreateAuthorizationResponseInfo:
    type: "object"
    properties:
      authorizationURL:
        type: "string"
        description: "The cloud storage URL to which the end-user should be directed\
          \ in order to authorize access by the \tCEC application."
      fieldValueMap:
        $ref: "#/definitions/CreateAuthorizationResponseInfo_fieldValueMap"
    description: "Information about the authorization-request."
    example:
      authorizationURL: "authorizationURL"
      fieldValueMap:
        name: "name"
        value: "value"
  CompleteAuthorizationRequestInfo:
    type: "object"
    properties:
      redirectURL:
        type: "string"
        description: "The redirect URL to which the authorization service should redirect\
          \ when finished. This will be the same value that was passed in the associated\
          \ CreateAuthorizationRequestInfo."
    description: "Completed-authorization creation info."
    example:
      redirectURL: "redirectURL"
  CompleteAuthorizationResponseInfo:
    type: "object"
    properties:
      authorized:
        type: "boolean"
        description: "Whether or not the connector is now authorized. Set to true\
          \ if the end-user granted authorization to the connector, set to false if\
          \ authorization was denied."
      authorizedUserDisplayName:
        type: "string"
        description: "The authorized user's display name, if available. May be null."
      authorizedUserEmailAddress:
        type: "string"
        description: "The authorized user's email address, if available. May be null."
      authorizedUserPictureURL:
        type: "string"
        description: "The authorized user's picture URL, if available. May be null."
      fieldValueMap:
        $ref: "#/definitions/CreateAuthorizationResponseInfo_fieldValueMap"
    description: "Completed-authorization information."
    example:
      authorizedUserEmailAddress: "authorizedUserEmailAddress"
      authorized: true
      fieldValueMap:
        name: "name"
        value: "value"
      authorizedUserPictureURL: "authorizedUserPictureURL"
      authorizedUserDisplayName: "authorizedUserDisplayName"
...
  FileSystemRequestInfo:
    type: "object"
    properties:
      itemCount:
        type: "number"
        description: "The number of items to return. Defaults to -1[All]"
      itemStartRow:
        type: "number"
        description: "The row number at which to start returning data. Used for pagination.\
          \ Defaults to 0"
      itemsSortField:
        type: "string"
        description: "Field Name on which to sort the records, default will be file\
          \ name."
      itemsSortOrder:
        type: "string"
        description: "ASC/DESC for ascending or descending sort. Defaults to Asc."
      uri:
        type: "string"
        description: "The object is identified by the uri parameter. If the uri parameter\
          \ is not set or is empty, then the method should return information about\
          \ the root or default file system."
      queryText:
        type: "string"
        description: "query keywords - search criteria."
    description: "File system information request information."
    example:
      itemsSortOrder: "itemsSortOrder"
      queryText: "queryText"
      itemStartRow: 6.02745618307040320615897144307382404804229736328125
      itemsSortField: "itemsSortField"
      uri: "uri"
      itemCount: 0.80082819046101150206595775671303272247314453125
  FileSystemResponseInfo:
    type: "object"
    properties:
      numItems:
        type: "number"
        description: "The number of items returned in response"
      hasMoreItems:
        type: "boolean"
        description: "This is true if the request did not return all of the item.\
          \ This occurs when itemsCount is reached and there are additional items\
          \ that could have been returned."
      totalItemsCount:
        type: "number"
        description: "The total number of items available."
      fileSystemInfo:
        $ref: "#/definitions/FilesystemInfo"
    description: "File system information response Object."
    example:
      totalItemsCount: 6.02745618307040320615897144307382404804229736328125
      hasMoreItems: true
      numItems: 0.80082819046101150206595775671303272247314453125
      fileSystemInfo:
        lastModifiedTimeStamp: 5.63737665663332876420099637471139430999755859375
        isThumbnailUrlPrivate: true
        additionalInformation:
          name: "name"
          value: "value"
        extension: "extension"
        creator: "creator"
        parentUri: "parentUri"
        browseURL: "browseURL"
        lastModifiedBy: "lastModifiedBy"
        createdTimeStamp: 5.962133916683182377482808078639209270477294921875
        directoryContents:
        - null
        - null
        description: "description"
        mimeType: "mimeType"
        uri: "uri"
        version: "version"
        size: 1.46581298050294517310021547018550336360931396484375
        name: "name"
        isDirectory: true
        thumbnailURL: "thumbnailURL"
  FilesystemInfo:
    type: "object"
    properties:
      uri:
        type: "string"
        description: "Uniform resource identifier for the file system object. This\
          \ identifier is used internally between the connector client and server\
          \ and is not intended to be used by the end-user. The format of this URI\
          \ is up to the connector. For example, it could be a numeric object ID,\
          \ a pathname, or a URL. The client will use this URI to identify the file\
          \ system object in other operations."
      parentUri:
        type: "string"
        description: "Uniform resource identifier for the parent file system object.\
          \ This identifier is used internally between the connector client and server\
          \ and is not intended to be used by the end-user. The format of this URI\
          \ is up to the connector. For example, it could be a numeric object ID,\
          \ a pathname, or a URL. The client will use this URI to identify the file\
          \ system object in other operations."
      name:
        type: "string"
        description: "The name of the file system object. This could be a directory\
          \ name or a file name (not the full pathname). This name is intended for\
          \ the end-user."
      description:
        type: "string"
        description: "Description of the file.  May be unset if the description is\
          \ unknown."
      isDirectory:
        type: "boolean"
        description: "Whether or not the file system object a directory."
      size:
        type: "number"
        description: "The size of the file in bytes. May be null if the file size\
          \ is unknown or not applicable to the file system object. This will be ignored\
          \ if <code>isDirectory</code> is true."
      mimeType:
        type: "string"
        description: "The MIME type of the file. May be unset if the MIME type is\
          \ unknown."
      extension:
        type: "string"
        description: "File extension. May be unset if the extension is unknown."
      creator:
        type: "string"
        description: "The original author of the file. May be unset if it unknown."
      createdTimeStamp:
        type: "number"
        description: "The created time of the file or directory in Unix time format\
          \ (seconds since the unix epoch; NOT Java milliseconds)."
      lastModifiedBy:
        type: "string"
        description: "User name who last modified the file. May be unset if it unknown."
      lastModifiedTimeStamp:
        type: "number"
        description: "The last modification time of the file or directory in Unix\
          \ time format (seconds since the unix epoch; NOT Java milliseconds). Do\
          \ not set this parameter if last modified user information is not available"
      browseURL:
        type: "string"
        description: "URL for the end-user to browse the file system object. This\
          \ is normally a URL  to the service provider when the user can view or edit\
          \ the contents of a file or browse the contents of a directory. Leave unset\
          \ if there is no such URL for the file system object."
      isThumbnailUrlPrivate:
        type: "boolean"
        description: "True : If thumbnail of a file will get fetch via connector otherwise\
          \ false"
      thumbnailURL:
        type: "string"
        description: "If isDirectory, the immediate contents of the directory. This\
          \ should not recurse beyond the level of file system objects contained in\
          \ this directory."
      version:
        type: "string"
        description: "Version of the file system object."
      directoryContents:
        type: "array"
        description: "If isDirectory, the immediate contents of the directory. This\
          \ should not recurse beyond the level of file system objects contained in\
          \ this directory."
        items:
          $ref: "#/definitions/FilesystemInfo"
      additionalInformation:
        $ref: "#/definitions/FilesystemInfo_additionalInformation"
    description: "Information about a file system object"
    example:
      lastModifiedTimeStamp: 5.63737665663332876420099637471139430999755859375
      isThumbnailUrlPrivate: true
      additionalInformation:
        name: "name"
        value: "value"
      extension: "extension"
      creator: "creator"
      parentUri: "parentUri"
      browseURL: "browseURL"
      lastModifiedBy: "lastModifiedBy"
      createdTimeStamp: 5.962133916683182377482808078639209270477294921875
      directoryContents:
      - null
      - null
      description: "description"
      mimeType: "mimeType"
      uri: "uri"
      version: "version"
      size: 1.46581298050294517310021547018550336360931396484375
      name: "name"
      isDirectory: true
      thumbnailURL: "thumbnailURL"
  CreateAuthorizationResponseInfo_fieldValueMap:
    properties:
      name:
        type: "string"
        description: "The property name. Required."
      value:
        type: "string"
        description: "The property value."
    description: "Any updates to the connector-settable custom fields."
    example:
      name: "name"
      value: "value"
  FilesystemInfo_additionalInformation:
    properties:
      name:
        type: "string"
        description: "The property name. Required."
      value:
        type: "string"
        description: "The property value."
    description: "Apart from regular meta data if connector wants to set additional\
      \ information."
    example:
      name: "name"
      value: "value"